requirements.txt
Pillow
ebooklib
natsort
